refactor onSelectDate
